[feature] - add new test elements update to DSL
- fix path lookup for factory tests
